Updated mnemonic.c
This test only reads the lower 16-bits of the instruction as that is probably
where instruction bits will be located with the upper 16-bits being used for
registers and immediates. Doing only the lower 16-bits also ensures that we
can actually test all of our combinations in a sane amount of time.