encodeBits now masks the value to encode into the input
I found a case where calls to encode-bits with different args yields the same
result:
    (encode-bits (input 0x000000FF)
                 (value 0x000F00FF)
                 (mask 0x0000FF00)
                 (shift 8))
    (encode-bits (input 0x000000FF)
                 (value 0x000F00FF)
                 (mask 0x0F00FF00)
                 (shift 8))
This is not supposed to happen so the cast and mask portion of encode is then
followed up with an and operation with the bitmask to make sure that only the
data desired is added. This was never encountered before because I was super
duper careful in the past. However, while implmenting the operation in iris
lisp I found this hole!