Added more simple-macros and defined the function call passing scheme
- Load/stores for direct/indirect persuasions of different widths have been
  defined as macros
- Catch system-call facts which occupy the same address and those which have a
  number of arguments not equal 0, 1, 2, or 3.

The function call passing scheme is now defined as a pointer in r6 which
contains at most 8 arguments stored in memory starting at the given address.
The arguments are loaded from memory as needed. NOTE that this could be a stack
address section or something similar, the use of r6 as this pointer makes this
concept very powerful. r7 is still used as the result return register