Added support for knowing the sizes of address and words for different cores and mmus
Since CLIPS uses 64-bit addresses, we have to make sure that we don't
waste space or have a consistent memory layout regardless of the core
accessing the rest of the system. So for the 16-bit iris cores, we have
to carve up a single memory address in the mmu's view into four separate
because of the word size. This applies to the cisc0 core as well.

this is the first step to allowing for synthetic computer system which
is controlled by the mmu (although this is just a conceptual thing).